:sparkles: すでに表示したページは子階層のページおよびカードに表示しない
